Add flexbox to .form-row and its child elements
For paragraph elements that are in the same .form-row div to sit next
to each other, the flexbox layout was applied to the .form-row div.

And for the elements inside each paragraph element to stack vertically,
the flexbox layout was applied along with the flex-direction property.

Lastly, a margin bottom was added to .form-row divs to provide space
between each .form-row div.